+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   playbooksSubfolder, a subfolder path to launch ansible-playbook from
+
 ## [6.0.0] - 2025-10-02
 
 ### Fixed